refactor: 클래스의 기능에 맞춰 package 분할
