+++ src/main/java/lotto/Application.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.domain.Machine;
+import lotto.controller.Machine;
 
 public class Application {
     public static void main(String[] args) {
