+++ src/main/java/lotto/service/Calculator.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.service;
+
+import lotto.domain.Rank;
 
 import java.util.List;
