+++ src/main/java/lotto/service/CompareLotto.java
@@ -1,4 +1,7 @@
-package lotto.domain;
+package lotto.service;
+
+import lotto.domain.Lotto;
+import lotto.domain.Rank;
 
 import java.util.List;
